Update README.md
Growth Mindset to mindset to me means staying positive and realizing that everything that is hard is just a process to learn and grow from trial and error. Three things for me to remember is to stay focus on the task and remember why I wanted to start pivoting in the first place. Secondly, I will try my very best and give 110 percent everyday so that I can reach my goals which is graduating. lastly I will be assertive in any way possible to be creative and look for answers the best way I can without asking for the answer.